go: separte type helper definition into type.go
